working on extract and use
